added dependencies properly. and changes to the data source.

+++ Budgie/Data Sources/WishListItemsDataSource.h
@@ -16,5 +16,5 @@
 
 -(void) parseItemsListWithCompletion:(void(^)(BOOL))completionBlock;
 
--(void) addItemWithIndex:(int)index toWishListWithCompletion:(void(^)(BOOL))completionBlock ;
+-(void) addItemWithIndex:(long)index toWishListWithCompletion:(void(^)(BOOL))completionBlock ;
 @end
